restore ctest cpp files while wating for implemented gtest
