+++ test/immed_conv3d.cpp
@@ -23,6 +23,6 @@
  * SOFTWARE.
  *
  *******************************************************************************/
-#include "conv3d.hpp"
+#include "immed_conv3d.hpp"
 
 int main(int argc, const char* argv[]) { test_drive<conv3d_driver>(argc, argv); }
