+++ test/immed_conv3d.hpp
@@ -23,13 +23,12 @@
  * SOFTWARE.
  *
  *******************************************************************************/
-#include <gtest/gtest.h>
 #include "conv_common.hpp"
 
 template <class T>
-struct conv3d_driver : conv_driver<T>
+struct conv3d_driver : conv_driver<T, ConvApi::Immediate>
 {
-    conv3d_driver() : conv_driver<T>()
+    conv3d_driver() : conv_driver<T, ConvApi::Immediate>()
     {
         this->add(this->input_dims, "input");
         this->add(this->weight_tensor_dims, "weights");
@@ -38,16 +37,16 @@ struct conv3d_driver : conv_driver<T>
                   this->generate_data_limited(this->get_batch_sizes(), 1, {8}));
         this->add(this->input_channels,
                   "input_channels",
-                  this->generate_data_limited(this->get_input_channels(), 1, {32}));
+                  this->generate_data_limited(this->get_input_channels(), 1, {2}));
         this->add(this->output_channels,
                   "output_channels",
-                  this->generate_data_limited(this->get_output_channels(), 1, {32}));
+                  this->generate_data_limited(this->get_output_channels(), 1, {16}));
         this->add(this->spatial_dim_elements,
                   "spatial_dim_elements",
                   this->generate_data_limited(this->get_3d_spatial_dims(), 1, {16, 16, 16}));
         this->add(this->filter_dims,
                   "filter_dims",
-                  this->generate_data_limited(this->get_3d_filter_dims(), 2, {3, 3, 3}));
+                  this->generate_data_limited(this->get_3d_filter_dims(), 2, {5, 5, 5}));
         this->add(this->pads_strides_dilations,
                   "pads_strides_dilations",
                   this->generate_data_limited(this->get_3d_pads_strides_dilations(), 2));
